centralize serialization into BaseConfig function
