+++ cumulusci/core/config/base_config.py
@@ -1,13 +1,22 @@
+import json
 import logging
 import os
 import typing as T
 import warnings
+from datetime import date, datetime
 from functools import lru_cache
 
 # turn on strictness.
 # delete this when strictness is mandatory
 STRICT_GETATTR = os.environ.get("STRICT_GETATTR")
 CHECK_CONFIG_TYPES = os.environ.get("CHECK_CONFIG_TYPES")
+
+
+def load_dates(x):
+    if isinstance(x, datetime):
+        return {"$type": "datetime", "$value": x.isoformat()}
+    elif isinstance(x, date):
+        return {"$type": "date", "$value": x.isoformat()}
 
 
 class BaseConfig(object):
@@ -42,6 +51,9 @@ class BaseConfig(object):
     def _load_config(self):
         """Subclasses may override this method to initialize :py:attr:`~config`"""
         pass
+
+    def _serialize(self):
+        return json.dumps(self.config, default=load_dates).encode("utf-8")
 
     @classmethod
     def _allowed_names(cls) -> T.Dict[str, type]:
